pull validation into own module
